Fail the clone operation when hg update fails.
There is a patch to update fix this issue for v1.40+ here:
https://github.com/jenkinsci/mercurial-plugin/pull/22

However v1.40+ is too new for the Khan build server (Jenkins v1.409).
This is a backport against v1.38.

Without this fix, when a subrepo fails to clone the last revision is
tracked as the nullid (40 zeros), which means that a subsequent email to
all "culprits" includes everyone who has ever committed.